Additions: (1) Added an NPC5Card class and some associated member functions
Changes:
(1) Moved and added some functionality in Player class hierarchy. Nothing is set in stone, plan is to keep things fluid and determine final locations as project nears completion.
(2) Five Card Draw Game: Added functions for game loop. Began working on said functions. Wrote buyInRound and dealHands, began bettingRound. bettingRound still needs dynamic betting for NPCs and for User.

Deletions:
(1) Deleted data member in DeckOfCards associated with high Aces as all Aces are now considered high. Also deleted any associated member functions.

+++ Card Games/Hand5.h
@@ -33,7 +33,7 @@ public:
 
 	//Public Member Functions
 	void addCard(const PlayingCard& card);
-	int count() const;
+	unsigned short int count() const;
 	void printHandStrings() const;
 	void printHandArt() const;
 	void replaceCardAt(size_t index, const PlayingCard& newCard);
